changed layout of the overview sidebar
