+++ client/src/assets/styles.css
@@ -115,7 +115,7 @@ select {
 
 .overview-sidebar {
   display: inline-block;
-  max-width: 500px;
+  width: 500px;
   padding: 0 20px;
 }
