+++ client/src/components/Overview/ProductInfo2.jsx
@@ -1,23 +1,6 @@
 import React from 'react';
-import StarRating from '../Ratings&Reviews/Ratings/StarRating.jsx';
 
-const ProductInfo = ({currentProduct, currentStyle, rating, totalReviews}) => {
-  const renderPrice = () => {
-    if (currentStyle?.sale_price) {
-      return (
-        <p>
-          <span style={{textDecoration: 'line-through red'}}>{currentStyle?.original_price}</span>
-          <span> </span>
-          <span style={{color: 'red'}}>{currentStyle?.sale_price}</span>
-        </p>
-      );
-    } else {
-      return (
-        <p>{currentStyle?.original_price}</p>
-      );
-    }
-  };
-
+const ProductInfo2 = ({currentProduct}) => {
   const onShare = (event) => {
     if (event.target.name === 'facebook') {
       window.open('https://facebook.com/');
@@ -30,11 +13,6 @@ const ProductInfo = ({currentProduct, currentStyle, rating, totalReviews}) => {
 
   return (
     <div>
-      {rating ? <StarRating rating={rating}/> : null}
-      {totalReviews ? <a href="#rateAndReview">Read all {totalReviews} reviews</a> : null}
-      <p>{currentProduct?.category?.toUpperCase()}</p>
-      <h1>{currentProduct?.name}</h1>
-      {renderPrice()}
       <p><i>{currentProduct?.slogan}</i></p>
       <p>{currentProduct?.description}</p>
       <p>Features</p>
@@ -50,4 +28,4 @@ const ProductInfo = ({currentProduct, currentStyle, rating, totalReviews}) => {
   );
 };
 
-export default ProductInfo;
+export default ProductInfo2;
